data-generator/requirements.txt
kafka-python
websocket-client